requirements.txt
fastapi~=0.111.1
redis~=5.0.8
starlette~=0.37.2
celery~=5.4.0
PyYAML~=6.0.1
python-dotenv~=1.0.1
SQLAlchemy~=2.0.31
pydantic~=2.8.2
aiohttp~=3.10.0
Brotli~=1.1.0
colorama~=0.4.6
pytest~=8.3.2
alembic~=1.13.2
uvicorn~=0.30.6
gunicorn~=23.0.0
asyncpg~=0.29.0
aiohttp-proxy~=0.1.2
aiohttp-socks~=0.8.4
flower~=2.0.1